+++ src/landscape.hpp
@@ -15,7 +15,7 @@ enum VegetationType {
 static_assert( sizeof(VegetationType) == 1 );
 
 struct Cell {
-  short elevation;
+  float elevation;
   float wind_direction;
   bool burnable;
   VegetationType vegetation_type;
@@ -26,7 +26,7 @@ struct Cell {
 struct Landscape {
   size_t width;
   size_t height;
-  Matrix<short> elevations;
+  Matrix<float> elevations;
   Matrix<float> wind_directions;
   Matrix<bool> burnables;
   Matrix<VegetationType> vegetation_types;
